move the gl3 debug stuff out of the bridge apply to all plats
