+++ ENIGMAsystem/SHELL/Bridges/SDL-OpenGL/graphics_bridge.cpp
@@ -31,6 +31,9 @@ int msaa_fbo = 0;
 SDL_GLContext context;
 
 void init_sdl_window_bridge_attributes() {
+  #ifdef DEBUG_MODE
+  SDL_GL_SetAttribute(SDL_GL_CONTEXT_FLAGS, SDL_GL_CONTEXT_DEBUG_FLAG);
+  #endif
   SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK,
     graphics_opengl_core?SDL_GL_CONTEXT_PROFILE_CORE:SDL_GL_CONTEXT_PROFILE_COMPATIBILITY);
   SDL_GL_SetSwapInterval(0);
