+++ ENIGMAsystem/SHELL/Bridges/Win32-OpenGL/graphics_bridge.cpp
@@ -32,50 +32,6 @@ namespace enigma {
 GLuint msaa_fbo = 0;
 extern HGLRC hRC;
 
-#ifdef DEBUG_MODE
-#include "Widget_Systems/widgets_mandatory.h"
-//Based on code from Cort Stratton (http://www.altdev.co/2011/06/23/improving-opengl-error-messages/)
-void FormatDebugOutputARB(char outStr[], size_t outStrSize, GLenum source, GLenum type, GLuint id, GLenum severity, const char *msg) {
-  char sourceStr[32]; const char *sourceFmt = "UNDEFINED(0x%04X)";
-  switch(source) {
-    case GL_DEBUG_SOURCE_API_ARB: sourceFmt = "API"; break;
-    case GL_DEBUG_SOURCE_WINDOW_SYSTEM_ARB: sourceFmt = "WINDOW_SYSTEM"; break;
-    case GL_DEBUG_SOURCE_SHADER_COMPILER_ARB: sourceFmt = "SHADER_COMPILER"; break;
-    case GL_DEBUG_SOURCE_THIRD_PARTY_ARB: sourceFmt = "THIRD_PARTY"; break;
-    case GL_DEBUG_SOURCE_APPLICATION_ARB: sourceFmt = "APPLICATION"; break;
-    case GL_DEBUG_SOURCE_OTHER_ARB: sourceFmt = "OTHER"; break;
-  }
-  snprintf(sourceStr, 32, sourceFmt, source);
-  char typeStr[32];
-  const char *typeFmt = "UNDEFINED(0x%04X)";
-  switch(type) {
-    case GL_DEBUG_TYPE_ERROR_ARB: typeFmt = "ERROR"; break;
-    case GL_DEBUG_TYPE_DEPRECATED_BEHAVIOR_ARB: typeFmt = "DEPRECATED_BEHAVIOR"; break;
-    case GL_DEBUG_TYPE_UNDEFINED_BEHAVIOR_ARB: typeFmt = "UNDEFINED_BEHAVIOR"; break;
-    case GL_DEBUG_TYPE_PORTABILITY_ARB: typeFmt = "PORTABILITY"; break;
-    case GL_DEBUG_TYPE_PERFORMANCE_ARB: typeFmt = "PERFORMANCE"; break;
-    case GL_DEBUG_TYPE_OTHER_ARB: typeFmt = "OTHER"; break;
-  }
-  snprintf(typeStr, 32, typeFmt, type);
-  char severityStr[32];
-  const char *severityFmt = "UNDEFINED(%i)";
-  switch(severity) {
-    case GL_DEBUG_SEVERITY_HIGH_ARB: severityFmt = "HIGH"; break;
-    case GL_DEBUG_SEVERITY_MEDIUM_ARB: severityFmt = "MEDIUM"; break;
-    case GL_DEBUG_SEVERITY_LOW_ARB: severityFmt = "LOW"; break;
-  }
-  snprintf(severityStr, 32, severityFmt, severity);
-  snprintf(outStr, outStrSize, "OpenGL: %s [source=%s type=%s severity=%s id=%d]", msg, sourceStr, typeStr, severityStr, id);
-}
-
-void DebugCallbackARB(GLenum source, GLenum type, GLuint id, GLenum severity, GLsizei length, const GLchar *message, GLvoid* userParam) {
-  char finalMessage[256];
-  FormatDebugOutputARB(finalMessage, 256, source, type, id, severity, message);
-  printf("%s\n", finalMessage);
-  show_error(toString(finalMessage), false);
-}
-#endif
-
 extern void (*WindowResizedCallback)();
 void WindowResized() {
   // clear the window color, viewport does not need set because backbuffer was just recreated
@@ -124,8 +80,6 @@ void EnableDrawing(void*)
       WGL_CONTEXT_PROFILE_MASK_ARB, graphics_opengl_core?WGL_CONTEXT_CORE_PROFILE_BIT_ARB:WGL_CONTEXT_COMPATIBILITY_PROFILE_BIT_ARB,
       #ifdef DEBUG_MODE
         WGL_CONTEXT_FLAGS_ARB, WGL_CONTEXT_DEBUG_BIT_ARB,
-      #else
-        WGL_CONTEXT_FLAGS_ARB, 0,
       #endif
       0
     };
@@ -134,24 +88,9 @@ void EnableDrawing(void*)
     wglMakeCurrent( NULL,NULL );
     wglDeleteContext( LegacyRC );
     wglMakeCurrent(enigma::window_hDC, hRC );
-  }
-  else // unable to get a core context, use the legacy context
-  {
+  } else { // unable to get a core context, use the legacy context
     hRC = LegacyRC;
   }
-
-  #ifdef DEBUG_MODE
-  glDebugMessageCallbackARB((GLDEBUGPROCARB)&DebugCallbackARB, 0);
-  glEnable(GL_DEBUG_OUTPUT_SYNCHRONOUS_ARB);
-  printf("OpenGL version supported by this platform (%s): \n", glGetString(GL_VERSION));
-
-  GLuint other_ids[] = { 131185 };
-  glDebugMessageControlARB(GL_DEBUG_SOURCE_API_ARB, GL_DEBUG_TYPE_OTHER_ARB, GL_DONT_CARE, 1, other_ids, GL_FALSE); //Disable some notifications shown below:
-  //OpenGL: Buffer detailed info: Buffer object 1 (bound to GL_ELEMENT_ARRAY_BUFFER_ARB, usage hint is GL_STATIC_DRAW) will use VIDEO memory as the source for buffer object operations. [source=API type=OTHER severity=UNDEFINED (33387) id=131185]
-  GLuint performance_ids[] = { 131218, 2 };
-  glDebugMessageControlARB(GL_DEBUG_SOURCE_API_ARB, GL_DEBUG_TYPE_PERFORMANCE_ARB, GL_DONT_CARE, 2, performance_ids, GL_FALSE); //Disable some notifications shown below:
-  //OpenGL: Program/shader state performance warning: Vertex shader in program 9 is being recompiled based on GL state. [source=API type=PERFORMANCE severity=MEDIUM id=131218] - This is NVidia only and doesn't tell much
-  #endif
 
   //TODO: This never reports higher than 8, but display_aa should be 14 if 2,4,and 8 are supported and 8 only when only 8 is supported
   glGetIntegerv(GL_MAX_SAMPLES_EXT, &enigma_user::display_aa);
